Teaching repo how to work with git clone
This patch gives repo the capability to use git clone instead of
git fetch. This makes it possible to take advantage of the git clone
--reference feature that makes it possible to sync a lot faster
than with fetch.

+++ git_config.py
@@ -257,9 +257,11 @@ class GitConfig(object):
       finally:
         fd.close()
     except IOError:
-      os.remove(self._pickle)
+      if os.path.exists(self._pickle):
+        os.remove(self._pickle)
     except cPickle.PickleError:
-      os.remove(self._pickle)
+      if os.path.exists(self._pickle):
+        os.remove(self._pickle)
 
   def _ReadGit(self):
     """
